feat(core): domain-free core step 4 — CLI commands + tool_handlers extraction (#885)
Step 4 of 8 in the domain-free-core refactor (see
docs/architecture/domain-free-core-audit.md §6).

Move ~370 LOC of game_ip-specific symbols out of core/cli/:
- core/cli/commands.py: cmd_list, cmd_generate, cmd_batch +
  14 game_ip slash entries in COMMAND_MAP + IP-specific show_help
  lines + Berserk example string → plugins/game_ip/cli/commands.py
- core/cli/tool_handlers.py: _build_analysis_handlers (180 LOC,
  6 handlers) + handle_generate_data + 4 signal entries from
  _DELEGATED_TOOLS (youtube/reddit/steam/google_trends) + FIXTURE_MAP
  reads → plugins/game_ip/cli/tool_handlers.py

DomainPort v2 (3 new methods, all with no-op defaults via getattr+callable):
  - get_rerunnable_nodes() -> set[str]: domain-supplied allowlist
    for /rerun_node (replaces hardcoded {"scoring","verification",
    "synthesizer"})
  - register_slash_commands(command_map): merge plugin slashes into
    COMMAND_MAP
  - register_tool_handlers(handlers): merge plugin handlers into
    the runtime handler dict

Wiring helpers:
  - core/cli/commands.py:install_domain_commands(domain) — bootstrap
    hook, mutates COMMAND_MAP once at startup
  - core/cli/tool_handlers.py:install_domain_tool_handlers(handlers)
    — runs at handler-build time inside _build_tool_handlers (closure
    context for verbose/force_dry/skill_registry)
  - plugins/game_ip/__init__.py: eager COMMAND_MAP merge at plugin
    import for static-import test contexts that don't run bootstrap

handle_rerun_node allowlist now sourced from domain.get_rerunnable_nodes().
handle_check_status fixture_count from domain.list_fixtures().
handle_show_help commands list from COMMAND_MAP.keys() (was hardcoded).

force_dry parameter on _build_analysis_handlers extended to bool|None
so the plugin can construct handlers without a build-time readiness
context (lazy lookup via _get_readiness when None).

No TID251 entries this step — core/cli/{commands,tool_handlers}.py
both still exist (just lighter), so module-level bans don't apply.
Step 5 onward will resume adding entries when whole modules move.

Verification:
- ruff/format/mypy/deptry/codespell: clean
- pytest -m "not live": 4386 passed, 1 skipped, 24 deselected
- E2E `geode analyze "Cowboy Bebop" --dry-run`: A (68.4) unchanged
- mypy source-file count: 243 → 245 (+2 new plugin modules)
