+++ CHANGELOG.md
@@ -28,6 +28,9 @@ Versioning follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Architecture
+- **Domain-free core, step 4 of 8 (`docs/architecture/domain-free-core-audit.md`).** CLI commands and tool-handler IP halves split out of `core/cli/`. `core/cli/commands.py` lost `cmd_list`, `cmd_generate`, `cmd_batch`, the 14 game-IP slash entries in `COMMAND_MAP` (`/analyze`, `/run`, `/list`, `/search`, `/report`, `/batch`, `/compare`, `/generate` + their aliases), and the IP examples block in `show_help`; the slashes now live in `plugins/game_ip/cli/commands.py:GAME_IP_SLASHES` and merge back into the generic `COMMAND_MAP` at bootstrap via the new `install_domain_commands(domain)` helper. `core/cli/tool_handlers.py` lost `_build_analysis_handlers` (180 LOC: `handle_list_ips` / `handle_analyze_ip` / `handle_search_ips` / `handle_compare_ips` / `handle_generate_report` / `handle_batch_analyze`), `handle_generate_data` (in `_build_execution_handlers`), the four signal entries in `_DELEGATED_TOOLS` (`youtube_search`, `reddit_sentiment`, `steam_info`, `google_trends`), and the `FIXTURE_MAP` reads in `handle_check_status` / `handle_show_help`; those handlers now live in `plugins/game_ip/cli/tool_handlers.py:build_game_ip_handlers()` and merge into the dispatcher dict at handler-build time via the new `install_domain_tool_handlers(handlers)` helper. The `handle_rerun_node` allowlist (`{"scoring", "verification", "synthesizer"}`) is now sourced from `domain.get_rerunnable_nodes()` instead of being hardcoded in core. `handle_check_status`'s `fixture_count` is now sourced from `domain.list_fixtures()`. `show_help` defers the IP-specific block to `domain.render_help_fragment()`. Three new optional `DomainPort` v2 methods declared in `core/domains/port.py` and implemented in `plugins/game_ip/adapter.py`: `get_rerunnable_nodes()`, `register_slash_commands(command_map)`, `register_tool_handlers(handlers)` — all use lazy plugin imports inside each method to avoid circular-import risk. `plugins/game_ip/__init__.py` also eagerly merges `GAME_IP_SLASHES` into `COMMAND_MAP` at import time so static-import paths (tests, REPL bootstrap, the legacy `COMMAND_REGISTRY` parity check) see the full slash registry without needing the bootstrap helper. After step 4, `grep -rn "from plugins\.game_ip\|import plugins\.game_ip" core/cli/{commands,tool_handlers}.py` returns zero hits except in `_handle_command` lazy imports for `/list`/`/generate`/`/batch` dispatch (which can't be removed until step 7-8 retire the `_handle_command` god method itself). `core/cli/routing.py` `/list` `handler_path` updated to `plugins.game_ip.cli.commands:cmd_list`. No TID251 entries land this step — step 4 only relocates symbols inside still-existing modules (`core/cli/commands.py` and `core/cli/tool_handlers.py` both shrink but stay), so module-level bans don't apply; resumes in step 5 when `core/tools/analysis.py` and the IP half of `core/tools/signal_tools.py` move whole-module. Quality gates green: ruff/mypy clean, deptry clean, 4386+ tests pass, E2E anchor `geode analyze "Cowboy Bebop" --dry-run` unchanged at A (68.4) undermarketed. (`core/domains/port.py`, `core/cli/{commands,tool_handlers,bootstrap,routing,__init__}.py`, `plugins/game_ip/__init__.py`, `plugins/game_ip/adapter.py`, `plugins/game_ip/cli/{commands,tool_handlers}.py`, `tests/test_commands.py`)
+
 ## [0.66.1] — 2026-05-06
 
 > **Hygiene + static analysis ratchet.** Post-v0.66.0 cleanup wave: 3
