+++ core/cli/routing.py
@@ -74,7 +74,7 @@ COMMAND_REGISTRY: dict[str, CommandSpec] = {
         name="/list",
         location=RunLocation.THIN,
         description="List available IPs",
-        handler_path="core.cli.commands:cmd_list",
+        handler_path="plugins.game_ip.cli.commands:cmd_list",
     ),
     "/login": CommandSpec(
         name="/login",
