+++ plugins/game_ip/cli/__init__.py
@@ -1,1 +1,1 @@
-"""Game IP plugin CLI helpers (batch, search, IP-name resolution)."""
+"""Game IP plugin CLI — slash commands, batch/search/IP-name resolution, tool handlers."""
